+++ framework/Furion/RemoteRequest/Extensions/RemoteRequestStringExtensions.cs
@@ -21,12 +21,13 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="requestUri"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> GetAsync(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> GetAsync(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Get, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, interceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Get, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, requestInterceptor, httpClientInterceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
         }
 
         /// <summary>
@@ -35,12 +36,13 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="requestUri"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> GetAsAsync<T>(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, CancellationToken cancellationToken = default)
+        public static async Task<T> GetAsAsync<T>(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Get, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, interceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Get, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, requestInterceptor, httpClientInterceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
         }
 
         /// <summary>
@@ -49,12 +51,13 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="requestUri"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> HeadAsync(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> HeadAsync(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Head, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, interceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Head, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, requestInterceptor, httpClientInterceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
         }
 
         /// <summary>
@@ -63,12 +66,13 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="requestUri"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> HeadAsAsync<T>(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, CancellationToken cancellationToken = default)
+        public static async Task<T> HeadAsAsync<T>(this string requestUri, Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Head, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, interceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Head, null, HttpContentTypeOptions.JsonStringContent, default, headers, clientName, requestInterceptor, httpClientInterceptor, JsonNamingPolicyOptions.CamelCase, cancellationToken);
         }
 
         /// <summary>
@@ -80,13 +84,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> PostAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> PostAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Post, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Post, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -98,13 +103,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> PostAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<T> PostAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Post, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Post, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -116,13 +122,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> PutAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> PutAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Put, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Put, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -134,13 +141,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> PutAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<T> PutAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Put, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Put, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -152,13 +160,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> DeleteAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> DeleteAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Delete, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Delete, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -170,13 +179,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> DeleteAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<T> DeleteAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Delete, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Delete, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -188,13 +198,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> PatchAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> PatchAsync(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsync(HttpMethod.Patch, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsync(HttpMethod.Patch, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -206,13 +217,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> PatchAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<T> PatchAsAsync<T>(this string requestUri, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            return await requestUri.SendAsAsync<T>(HttpMethod.Patch, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            return await requestUri.SendAsAsync<T>(HttpMethod.Patch, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
         }
 
         /// <summary>
@@ -225,13 +237,14 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<T> SendAsAsync<T>(this string requestUri, HttpMethod httpMethod = default, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<T> SendAsAsync<T>(this string requestUri, HttpMethod httpMethod = default, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
-            var response = await requestUri.SendAsync(httpMethod, bodyArgs, httpContentType, contentType, headers, clientName, interceptor, propertyNamingPolicy, cancellationToken);
+            var response = await requestUri.SendAsync(httpMethod, bodyArgs, httpContentType, contentType, headers, clientName, requestInterceptor, httpClientInterceptor, propertyNamingPolicy, cancellationToken);
 
             // 读取流
             using var responseStream = await response.Content.ReadAsStreamAsync(cancellationToken);
@@ -251,11 +264,12 @@ namespace Furion.RemoteRequest.Extensions
         /// <param name="contentType"></param>
         /// <param name="headers"></param>
         /// <param name="clientName"></param>
-        /// <param name="interceptor"></param>
+        /// <param name="requestInterceptor"></param>
+        /// <param name="httpClientInterceptor"></param>
         /// <param name="propertyNamingPolicy"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public static async Task<HttpResponseMessage> SendAsync(this string requestUri, HttpMethod httpMethod = default, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> interceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
+        public static async Task<HttpResponseMessage> SendAsync(this string requestUri, HttpMethod httpMethod = default, object bodyArgs = null, HttpContentTypeOptions httpContentType = HttpContentTypeOptions.JsonStringContent, string contentType = "application/json", Dictionary<string, string> headers = default, string clientName = default, Action<HttpRequestMessage> requestInterceptor = default, Action<HttpClient> httpClientInterceptor = default, JsonNamingPolicyOptions propertyNamingPolicy = JsonNamingPolicyOptions.CamelCase, CancellationToken cancellationToken = default)
         {
             // 检查 method
             if (httpMethod == null) throw new ArgumentNullException(nameof(httpMethod));
@@ -279,7 +293,7 @@ namespace Furion.RemoteRequest.Extensions
             }
 
             // 请求之前拦截
-            interceptor?.Invoke(request);
+            requestInterceptor?.Invoke(request);
 
             // 打印请求地址
             App.PrintToMiniProfiler(Penetrates.MiniProfilerCategory, "Beginning", $"{request.Method} {request.RequestUri}");
@@ -291,6 +305,9 @@ namespace Furion.RemoteRequest.Extensions
             var httpClient = string.IsNullOrEmpty(clientName)
                                         ? clientFactory.CreateClient()
                                         : clientFactory.CreateClient(clientName);
+
+            // 拦截客户端
+            httpClientInterceptor?.Invoke(httpClient);
 
             // 发送请求
             var response = await httpClient.SendAsync(request, cancellationToken);